app/requirements.txt
asgiref==3.5.2
Django==4.1.3
django-cors-headers==3.13.0
django-crontab==0.7.1
django-jalali-date==1.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
gunicorn==20.1.0
jdatetime==4.1.0
numpy==1.24.2
openpyxl==3.1.2
pandas==2.0.0
psycopg2==2.9.9
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
tzdata==2022.6
